add copyright notices to all files which may containes code that was carried over
