+++ api/agent/common/v1/common.proto
@@ -4,7 +4,6 @@ option go_package = "github.com/cisco-open/nasp/api/agent/common/v1;commonv1";
 
 import "google/protobuf/empty.proto";
 
-// Manages registration entries stored by the SPIRE Server.
 service Common {
     rpc ResetWASM(google.protobuf.Empty) returns (google.protobuf.Empty);
 }
